shared data.h file updated float calculation

+++ Headerfile/CurrentCalculation.h
@@ -23,5 +23,5 @@ void setup_currentMeasurementValues(void);
  * @param adc_value The ADC value representing the current measurement.
  * @return 1 if the current is within the acceptable range, 0 otherwise (indicating a failure).
  */
-float calculate_current(uint32_t adc_value);
+uint32_t calculate_current(uint32_t adc_value);
 #endif // CURRENT_MEASUREMENT_H
