+++ Headerfile/measurement.h
@@ -26,9 +26,9 @@ uint32_t adc_to_rpm(uint32_t adc_value, uint32_t max_rpm, uint32_t adc_max_value
  */
 uint32_t calculate_throttle(uint32_t adc_value, uint32_t max_rpm);
 
-float calculate_voltage(uint32_t adc_value);
-float calculate_current(uint32_t adc_value);
-float measure_temperature(uint32_t adc_value);
+uint32_t calculate_voltage(uint32_t adc_value);
+uint32_t calculate_current(uint32_t adc_value);
+uint32_t measure_temperature(uint32_t adc_value);
 void calculateMotorPeriod(uint32_t cap);
 void resetMotorPeriod(void);
 void filterMotorPeriod(void);
